GR-69858: Move hosted security provider support to JCA package
