+++ substratevm/docs/architecture/security-providers.md
@@ -1,1 +1,1 @@
-# AR-security-providers: [SecurityServicesFeature](../../src/com.oracle.svm.hosted/src/com/oracle/svm/hosted/SecurityServicesFeature.java)
+# AR-security-providers: [SecurityServicesFeature](../../src/com.oracle.svm.hosted/src/com/oracle/svm/hosted/jca/SecurityServicesFeature.java)
