+++ substratevm/src/com.oracle.graal.pointsto/src/com/oracle/graal/pointsto/meta/AnalysisUniverse.java
@@ -410,11 +410,6 @@ public class AnalysisUniverse implements Universe {
                         ". Probably there are some compilation or classpath problems. ");
     }
 
-    public AnalysisMethod optionalLookup(ResolvedJavaMethod method) {
-        ResolvedJavaMethod actualMethod = substitutions.lookup(method);
-        return methods.get(actualMethod);
-    }
-
     @Override
     public JavaMethod lookupAllowUnresolved(JavaMethod rawMethod) {
         if (rawMethod == null) {
