+++ substratevm/src/com.oracle.svm.hosted.test/src/com/oracle/svm/hosted/test/VerifyReflectionUsage.java
@@ -294,10 +294,10 @@ public class VerifyReflectionUsage extends VerifyReflectionUsageBase {
                     clazz("com.oracle.svm.hosted.ResourcesFeature$1"),
                     clazz("com.oracle.svm.hosted.ResourcesFeature$ResourceCollectorImpl"),
                     clazz("com.oracle.svm.hosted.ResourcesFeature$ResourcesRegistryImpl"),
+                    clazz("com.oracle.svm.hosted.jca.SecurityServicesFeature"),
                     clazz("com.oracle.svm.hosted.sboutlining.concat.SubstrateStringConcatFactory"),
                     clazz("com.oracle.svm.hosted.sboutlining.concat.SubstrateStringConcatGraphBuilder"),
                     clazz("com.oracle.svm.hosted.sboutlining.SBOutliningAnalysis$MetadataLookup"),
-                    clazz("com.oracle.svm.hosted.SecurityServicesFeature"),
                     clazz("com.oracle.svm.hosted.snippets.ReflectionPlugins"),
                     clazz("com.oracle.svm.hosted.snippets.ReflectionPlugins$3"),
                     clazz("com.oracle.svm.hosted.substitute.AutomaticUnsafeTransformationSupport"),
