+++ substratevm/src/com.oracle.svm.hosted/src/com/oracle/svm/hosted/jca/SecurityProviderCatalogRegistrar.java
@@ -22,7 +22,7 @@
  * or visit www.oracle.com if you need additional information or have any
  * questions.
  */
-package com.oracle.svm.hosted;
+package com.oracle.svm.hosted.jca;
 
 import java.security.Provider;
 import java.security.Provider.Service;
